Fix compile error with gcc 7.3, CUDA 9.2
Former-commit-id: a35a57db165056d4c3cb7db298a6804a5f521d5b

+++ src/Profile/CuptiActivity.cpp
@@ -1883,7 +1883,7 @@ ImixStats write_runtime_imix(uint32_t functionId, std::map<std::pair<int, int>, 
       if ( srcLocMap.find(sid) != srcLocMap.end() ) {
 	lineno = srcLocMap.find(sid)->second.lineNumber;
 	// cout << "[CuptiActivity]:  lineno: " << lineno << endl;
-	std::pair<int, int> p1 = std::make_pair(lineno, is.pcOffset);
+	std::pair<int, int> p1 = std::make_pair(lineno, (unsigned int) is.pcOffset);
 
 	for (std::map<std::pair<int, int>,CudaOps>::iterator iter= map_disassem.begin();
 	     iter != map_disassem.end(); iter++) {
